+++ src/features/Leaderboards/index.js
@@ -15,10 +15,22 @@ const Leaderboards = () => (
     <Tabs
       titleSuffix={T.translate('leaderboards.name')}
       tabs={[{
-        to: '/leaderboards',
-        name: 'Pvp (GW2A)',
-        content: <PvpLeaderboard />,
+        to: '/leaderboards/pvp',
+        name: 'PvP (GW2A)',
+        content: <PvpLeaderboard region="gw2a" />,
         description: config.descriptions.pvpLeaderboard,
+      }, {
+        to: '/leaderboards/pvp/na',
+        name: 'PvP (NA)',
+        content: <PvpLeaderboard region="na" />,
+        description: config.descriptions.pvpLeaderboard,
+        flair: 'new',
+      }, {
+        to: '/leaderboards/pvp/eu',
+        name: 'PvP (EU)',
+        content: <PvpLeaderboard region="eu" />,
+        description: config.descriptions.pvpLeaderboard,
+        flair: 'new',
       }]}
     />