+++ src/features/User/index.js
@@ -9,6 +9,7 @@ import filter from 'lodash/filter';
 import T from 'i18n-react';
 import { Link } from 'react-router';
 
+import Button from 'common/components/Button';
 import Icon from 'common/components/Icon';
 import Content from 'common/layouts/Content';
 import ContentCardList from 'common/components/ContentCardList';
@@ -56,6 +57,8 @@ function getActiveStanding ({ pvpStandings = [] } = {}, pvpSeasons) {
   return standings[0] || {};
 }
 
+const addHash = (str) => (str ? `#${str}` : '-');
+
 @connect(selector, {
   dispatchFetchUser: fetchUser,
   dispatchSelectUser: selectUser,
@@ -92,9 +95,12 @@ export default class User extends Component {
     const pvpGames = (get(user, 'pvpGames.length') && get(user, 'pvpGames')) || [undefined, undefined];
     const guilds = get(user, 'guilds', []);
 
+    const stubUser = get(user, 'stub');
     const standing = getActiveStanding(user, pvpSeasons);
-    const rating = get(standing, 'current.rating');
+    const rating = get(standing, 'current.rating') || get(user, 'rating');
     const gw2aRank = get(user, 'gw2aRank');
+    const euRank = get(user, 'euRank');
+    const naRank = get(user, 'naRank');
     const winsText = T.translate('users.pvpStats.wins');
     const byesText = T.translate('users.pvpStats.byes');
     const lossText = T.translate('users.pvpStats.losses');
@@ -111,8 +117,20 @@ export default class User extends Component {
         rightComponent={
           <ul className={styles.rating}>
             <li>
-              <Link to="/leaderboards">
-                <span>{(gw2aRank && `#${gw2aRank}`) || '-'}</span>
+              <Link to="/leaderboards/pvp/eu">
+                <span>{addHash(euRank)}</span>
+                <div>EU {T.translate('users.pvpStats.ranking')}</div>
+              </Link>
+            </li>
+            <li>
+              <Link to="/leaderboards/pvp/na">
+                <span>{addHash(naRank)}</span>
+                <div>NA {T.translate('users.pvpStats.ranking')}</div>
+              </Link>
+            </li>
+            <li>
+              <Link to="/leaderboards/pvp">
+                <span>{addHash(gw2aRank)}</span>
                 <div>GW2A {T.translate('users.pvpStats.ranking')}</div>
               </Link>
             </li>
@@ -129,6 +147,11 @@ export default class User extends Component {
         type="users"
         title={alias}
         content={user}
+        pinnedTab={stubUser && (
+          <Link to={this.context._userAlias ? `settings?claiming=${alias}` : `join?claiming=${alias}`}>
+            <Button type="cta">{T.translate('users.claimCta')}</Button>
+          </Link>
+        )}
         tabs={[
           {
             to: `/${alias}`,
@@ -153,7 +176,6 @@ export default class User extends Component {
           {
             to: `/${alias}/guilds`,
             name: T.translate('guilds.name'),
-            flair: 'new',
             content: (
               <ContentCardList
                 noBorder
@@ -170,7 +192,8 @@ export default class User extends Component {
             content: (
               <div className={styles.gamesContainer}>
                 <span />
-                {pvpGames.map((game, index) => <PvpGame game={game} key={game ? game.id : index} maps={maps} />)}
+                {pvpGames.map((game, index) =>
+                  <PvpGame game={game} key={game ? game.id : index} maps={maps} />)}
               </div>
             ),
           },